+++ sites/mip.39.net/components/mip-quick-router/example/index.html
@@ -12,7 +12,7 @@
   </head>
   <body>
     <div>
-      <mip-quick-router packagename="com.qisheng.hybrid" page="EnterPage" params="pid=56990211" ></mip-quick-router>
+      <mip-quick-router approuter='false' packagename="com.qisheng.hybrid" page="EnterPage" params="pid=56990211" ></mip-quick-router>
     </div>
     <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
     <script src="/mip-quick-router/mip-quick-router.js"></script>
